daemon: Move bootID load into core New()
Doesn't make sense to do this in `start()` and pass it through
the ctors, we only get one value.
