+++ cmd/machine-config-daemon/start.go
@@ -2,11 +2,9 @@ package main
 
 import (
 	"flag"
-	"io/ioutil"
 	"os"
 	"os/exec"
 	"path/filepath"
-	"strings"
 	"syscall"
 
 	"github.com/golang/glog"
@@ -46,15 +44,6 @@ func init() {
 	startCmd.PersistentFlags().BoolVar(&startOpts.skipReboot, "skip-reboot", false, "Skips reboot after a sync, applies only in once-from")
 	startCmd.PersistentFlags().BoolVar(&startOpts.kubeletHealthzEnabled, "kubelet-healthz-enabled", true, "kubelet healthz endpoint monitoring")
 	startCmd.PersistentFlags().StringVar(&startOpts.kubeletHealthzEndpoint, "kubelet-healthz-endpoint", "http://localhost:10248/healthz", "healthz endpoint to check health")
-}
-
-// getBootID loads the unique "boot id" which is generated by the Linux kernel.
-func getBootID() (string, error) {
-	currentBootIDBytes, err := ioutil.ReadFile("/proc/sys/kernel/random/boot_id")
-	if err != nil {
-		return "", err
-	}
-	return strings.TrimSpace(string(currentBootIDBytes)), nil
 }
 
 // bindPodMounts ensures that the daemon can still see e.g. /run/secrets/kubernetes.io
@@ -101,11 +90,6 @@ func runStartCmd(cmd *cobra.Command, args []string) {
 		glog.Fatalf("Error found when checking operating system: %s", err)
 	}
 
-	bootID, err := getBootID()
-	if err != nil {
-		glog.Fatalf("Cannot get boot ID: %v", err)
-	}
-
 	if startOpts.nodeName == "" {
 		name, ok := os.LookupEnv("NODE_NAME")
 		if !ok || name == "" {
@@ -137,7 +121,6 @@ func runStartCmd(cmd *cobra.Command, args []string) {
 			startOpts.nodeName,
 			operatingSystem,
 			daemon.NewNodeUpdaterClient(),
-			bootID,
 			startOpts.onceFrom,
 			startOpts.skipReboot,
 			nil,
@@ -179,7 +162,6 @@ func runStartCmd(cmd *cobra.Command, args []string) {
 		daemon.NewNodeUpdaterClient(),
 		ctx.InformerFactory.Machineconfiguration().V1().MachineConfigs(),
 		kubeClient,
-		bootID,
 		startOpts.onceFrom,
 		startOpts.skipReboot,
 		ctx.KubeInformerFactory.Core().V1().Nodes(),
