+++ pkg/daemon/daemon.go
@@ -171,13 +171,21 @@ func rebootCommand(rationale string) *exec.Cmd {
 		"--description", fmt.Sprintf("machine-config-daemon: %s", rationale), "/bin/sh", "-c", "systemctl stop kubelet.service; reboot")
 }
 
+// getBootID loads the unique "boot id" which is generated by the Linux kernel.
+func getBootID() (string, error) {
+	currentBootIDBytes, err := ioutil.ReadFile("/proc/sys/kernel/random/boot_id")
+	if err != nil {
+		return "", err
+	}
+	return strings.TrimSpace(string(currentBootIDBytes)), nil
+}
+
 // New sets up the systemd and kubernetes connections needed to update the
 // machine.
 func New(
 	nodeName,
 	operatingSystem string,
 	nodeUpdaterClient NodeUpdaterClient,
-	bootID,
 	onceFrom string,
 	skipReboot bool,
 	mcClient mcfgclientset.Interface,
@@ -200,6 +208,11 @@ func New(
 			return nil, fmt.Errorf("error reading osImageURL from rpm-ostree: %v", err)
 		}
 		glog.Infof("Booted osImageURL: %s (%s)", osImageURL, osVersion)
+	}
+
+	bootID, err := getBootID()
+	if err != nil {
+		return nil, errors.Wrapf(err, "failed to read boot ID")
 	}
 
 	dn := &Daemon{
@@ -233,7 +246,6 @@ func NewClusterDrivenDaemon(
 	nodeUpdaterClient NodeUpdaterClient,
 	mcInformer mcfginformersv1.MachineConfigInformer,
 	kubeClient kubernetes.Interface,
-	bootID,
 	onceFrom string,
 	skipReboot bool,
 	nodeInformer coreinformersv1.NodeInformer,
@@ -247,7 +259,6 @@ func NewClusterDrivenDaemon(
 		nodeName,
 		operatingSystem,
 		nodeUpdaterClient,
-		bootID,
 		onceFrom,
 		skipReboot,
 		nil,
